requirements.txt
# local package
-e .

# external requirements

Sphinx
coverage
flake8


click==8.1.3
matplotlib==3.6.2
numpy==1.24.1
Pillow==9.4.0
python-dotenv==0.21.0
setuptools==65.6.3
torch==1.13.0
torchvision
pytorch_lightning==1.8.1
torchmetrics==0.11.0
transformers==4.24.0


hydra-core==1.3.1







